Move CustomAutoSchema class to a dedicated tools_drf.py file to prevent a circular dependency.
