+++ admin_cohort/backends.py
@@ -1,7 +1,7 @@
 from typing import Union
 
 from admin_cohort import conf_auth
-from admin_cohort.example_conf_auth import LoginError, ServerError, JwtTokens
+from admin_cohort.conf_auth import LoginError, ServerError, JwtTokens
 from admin_cohort.models import User, get_or_create_user
