+++ admin_cohort/tools.py
@@ -5,7 +5,6 @@ from operator import or_
 from typing import List, Union
 
 from django.db.models import Q
-from drf_yasg.inspectors import SwaggerAutoSchema
 
 
 def prettify_dict(obj: Union[dict, list]) -> str:
@@ -19,17 +18,7 @@ def prettify_json(str_json: str) -> str:
         return str_json
 
 
-# seen on https://stackoverflow.com/a/64440802
-class CustomAutoSchema(SwaggerAutoSchema):
-    def get_tags(self, operation_keys=None):
-        tags = self.overrides.get('tags', None) \
-               or getattr(self.view, 'swagger_tags', [])
-        if not tags:
-            tags = [operation_keys[0]]
-
-        return tags
-# DJANGO Q HELPER
-
+# DJANGO Q HELPERS
 
 def join_qs(qs: List[Q]) -> Q:
     """
